pyproject.toml
[project]
name = "antihyip-rest"
version = "0.1.0"
description = "Website and API to help user identify investment scams, like HYIP-pyramids, etc."
authors = [
    { name = "Roman Krotov", email = "152505489+sishorian@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "django-guardian>=2.4.0",
    "djangorestframework>=3.16.0",
    "ipython>=9.2.0",
    "markdown>=3.8",
    "pygments>=2.19.1",
    "pyopenssl>=25.0.0",
    "pyyaml>=6.0.2",
    "uritemplate>=4.1.1",
    "werkzeug>=3.1.3",
]

[tool.djlint]
profile="django"
indent=2
max_line_length=88
preserve_blank_lines=true
max_blank_lines=1

[tool.ruff.lint]
ignore = [
    "E402", # module-import-not-at-top-of-file
    # Prefer pyright's diagnostics
    "F401", # unused-import
    "F821", # undefined name
    "F841", # unused-variable (assigned but not used)
    "F842", # unused-annotation (declaration)
]

[tool.basedpyright]
# Prefer Ruff's diagnostic
reportRedeclaration = "none"
# Django
typeCheckingMode = "standard"
reportAttributeAccessIssue = "none"
reportUnusedParameter = "none"
reportUnreachable = "none"
reportArgumentType = "none"
reportAssignmentType = "none"
reportIncompatibleMethodOverride = "none"
reportFunctionMemberAccess = "none"
